refactor: barrel-purity-checker.js を utils.js にリネームし、重複チェック関連の関数を移動
変更内容:
- barrel-purity-checker.js → utils.js にリネーム
- utils.js に以下の関数を追加:
  - extractFileName: パスからファイル名（拡張子なし）を抽出
  - generateBarrelFilePaths: barrelファイル名の候補パスを生成
  - getSiblingBarrelFiles: 同じディレクトリ内の他のバレルファイルを取得
  - extractExportsFromBarrel: バレルファイルからexport識別子を抽出
- index.js から上記の関数定義を削除し、utils.js からimport

理由:
重複exportチェック機能で追加した関数をユーティリティとして
utils.js に集約し、コードの保守性を向上させました。
