+++ packages/eslint-plugin-smarthr/rules/require-barrel-import/index.js
@@ -2,7 +2,14 @@ const path = require('path')
 const fs = require('fs')
 const { replacePaths, rootPath } = require('../../libs/common')
 const { getParentDir } = require('../../libs/util')
-const { isBarrelFile, createBarrelPurityVisitor } = require('./barrel-purity-checker')
+const {
+  isBarrelFile,
+  createBarrelPurityVisitor,
+  extractFileName,
+  generateBarrelFilePaths,
+  getSiblingBarrelFiles,
+  extractExportsFromBarrel,
+} = require('./utils')
 
 const SCHEMA = [
   {
@@ -112,76 +119,6 @@ const findCommonParent = (path1, path2) => {
   }
 
   return segments1.slice(0, i).join('/')
-}
-
-/**
- * 指定されたbarrelファイル名の候補パスを生成する
- * @param {string} dir - ディレクトリパス
- * @param {Array<string>} fileNames - barrelファイル名の配列
- * @returns {Array<string>} パス候補の配列
- */
-const generateBarrelFilePaths = (dir, fileNames) => {
-  return fileNames.flatMap(name => TARGET_EXTS.map(ext => `${dir}/${name}.${ext}`))
-}
-
-/**
- * パスからファイル名（拡張子なし）を抽出する
- * @param {string} filePath - ファイルパス
- * @returns {string} ファイル名（拡張子なし）
- */
-const extractFileName = (filePath) => {
-  return filePath.split('/').pop().replace(REGEX_BARREL_FILE_EXT, '')
-}
-
-/**
- * 同じディレクトリ内の他のバレルファイルを取得
- * @param {string} currentBarrelPath - 現在のバレルファイルの絶対パス
- * @param {Array<string>} barrelFileNames - バレルファイル名のリスト
- * @returns {Array<string>} 他のバレルファイルのパスの配列
- */
-const getSiblingBarrelFiles = (currentBarrelPath, barrelFileNames) => {
-  const dir = getParentDir(currentBarrelPath)
-  const currentFileName = extractFileName(currentBarrelPath)
-
-  return generateBarrelFilePaths(dir, barrelFileNames)
-    .filter(barrelPath => {
-      return barrelPath !== currentBarrelPath &&
-             extractFileName(barrelPath) !== currentFileName &&
-             fs.existsSync(barrelPath)
-    })
-}
-
-/**
- * バレルファイルからexportされている識別子を抽出
- * @param {string} barrelFilePath - バレルファイルのパス
- * @returns {Set<string>} exportされている識別子のSet
- */
-const extractExportsFromBarrel = (barrelFilePath) => {
-  const exports = new Set()
-
-  try {
-    const content = fs.readFileSync(barrelFilePath, 'utf-8')
-
-    // export { A, B } from './file' パターン
-    const namedExportPattern = /export\s*\{\s*([^}]+)\s*\}\s*from\s*['"]([^'"]+)['"]/g
-    let match
-    while ((match = namedExportPattern.exec(content)) !== null) {
-      const exportedNames = match[1].split(',').map(name => {
-        // "as" を使っている場合は、asの後の名前を取得
-        const asMatch = name.trim().match(/^(.+?)\s+as\s+(.+)$/)
-        return asMatch ? asMatch[2].trim() : name.trim()
-      })
-      exportedNames.forEach(name => exports.add(name))
-    }
-
-    // export * from './file' パターン（すべてをre-exportするので、重複チェック対象外とする）
-    // export * as namespace from './file' は含まない
-
-  } catch (err) {
-    // ファイル読み込みエラーは無視（存在チェックは呼び出し側で行う）
-  }
-
-  return exports
 }
 
 /**
